Models+Controllers+Views+DAO de Profissional

+++ app/dao/CargoDAO.java
@@ -119,7 +119,7 @@ public class CargoDAO {
 		return cargos;
 	}
 	
-	public MCargo getCargo(int id){
+	public MCargo getCargo(Long id){
 		
 		String sql = "select * from cargo where id = " + id;
